Refactoring and edit of objectives
