+++ src/client/actions.js
@@ -274,3 +274,5 @@ export const COMPANY_LISTS__COMPANY_IN_LOADED =
   'COMPANY_LISTS__COMPANY_IN_LOADED'
 
 export const DNB_FAMILY_TREE_LOADED = 'DNB_FAMILY_TREE_LOADED'
+
+export const OBJECTIVE_LOADED = 'OBJECTIVE_LOADED'
