+++ src/client/index.jsx
@@ -394,6 +394,9 @@ import { saveStrategy } from './modules/Companies/AccountManagement/tasks'
 import { TASK_SAVE_OBJECTIVE } from './modules/Companies/AccountManagement/state'
 import { saveObjective } from './modules/Companies/AccountManagement/tasks'
 
+import { TASK_GET_OBJECTIVE } from './modules/Companies/AccountManagement/state'
+import { getObjective } from './modules/Companies/AccountManagement/tasks'
+
 function parseProps(domNode) {
   return 'props' in domNode.dataset ? JSON.parse(domNode.dataset.props) : {}
 }
@@ -626,6 +629,7 @@ function App() {
           [TASK_SAVE_STRATEGY]: saveStrategy,
           [TASK_EDIT_INVESTMENT_PROJECT_STATUS]: updateInvestmentProject,
           [TASK_SAVE_OBJECTIVE]: saveObjective,
+          [TASK_GET_OBJECTIVE]: getObjective,
         }}
       >
         <Mount selector="#data-hub-header">
