+++ src/client/modules/Companies/AccountManagement/tasks.js
@@ -7,9 +7,21 @@ export const saveStrategy = ({ strategy, companyId }) => {
   return request(endpoint, { strategy: strategy })
 }
 
-export const saveObjective = ({ values, companyId }) => {
+export const saveObjective = ({ values, companyId, objectiveId }) => {
   values.company = companyId
-  const request = apiProxyAxios.post
-  const endpoint = `/v4/company/${companyId}/objective`
-  return request(endpoint, transformFormValuesForAPI(values))
+  // values.id = objectiveId
+  if (objectiveId) {
+    const request = apiProxyAxios.patch
+    const endpoint = `/v4/company/${companyId}/objective/${objectiveId}`
+    return request(endpoint, transformFormValuesForAPI(values))
+  } else {
+    const request = apiProxyAxios.post
+    const endpoint = `/v4/company/${companyId}/objective`
+    return request(endpoint, transformFormValuesForAPI(values))
+  }
 }
+
+export const getObjective = ({ companyId, objectiveId }) =>
+  apiProxyAxios
+    .get(`/v4/company/${companyId}/objective/${objectiveId}`)
+    .then(({ data }) => data)
